Use correct integer types in version

+++ src/sdl2/version.rs
@@ -11,11 +11,11 @@ pub use sys::version as ll;
 #[derive(PartialEq, Copy, Clone)]
 pub struct Version {
     /// major version
-    pub major: isize,
+    pub major: u8,
     /// minor version
-    pub minor: isize,
+    pub minor: u8,
     /// update version (patchlevel)
-    pub patch: isize,
+    pub patch: u8,
 }
 
 impl Version {
@@ -23,7 +23,7 @@ impl Version {
     pub fn from_ll(sv: *const ll::SDL_version) -> Version {
         unsafe {
             let ref v = *sv;
-            Version{ major: v.major as isize, minor: v.minor as isize, patch: v.patch as isize }
+            Version{ major: v.major, minor: v.minor, patch: v.patch }
         }
     }
 }
@@ -52,8 +52,8 @@ pub fn get_revision() -> String {
 }
 
 /// Get the revision number of SDL that is linked against your program.
-pub fn get_revision_number() -> isize {
+pub fn get_revision_number() -> i32 {
     unsafe {
-        ll::SDL_GetRevisionNumber() as isize
+        ll::SDL_GetRevisionNumber()
     }
 }
